Create Number Planet.cpp
Welcome to Planet Numero!
In this land, numbers are people. All the factors of a number are its descendants. A and B are two such people who meet at a party. As they get to know each other they realise that they are in fact related!!

In order to trace the common members of their family, they need your help. Identify the eldest descendent of both A and B and the youngest common ancestor they share.